+++ packages/app/src/interfaces/in-app-notification-settings.ts
@@ -1,0 +1,10 @@
+import { Schema } from 'mongoose';
+
+export interface IDefaultSubscribeRule {
+  name: string,
+  isEnabled: boolean;
+}
+export interface IInAppNotificationSettings {
+  userId: Schema.Types.ObjectId;
+  defaultSubscribeRules: IDefaultSubscribeRule[];
+}
